+++ src/AssemblyInfo.cs
@@ -1,0 +1,4 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("dec-test-unit")]
+[assembly: InternalsVisibleTo("fuzzgen")]